Improve type definition for state calculation params Refactor Use suggestion save for state calculation
